refactor: context 구조화
