+++ packages/assignment/src/context/AppProvider.tsx
@@ -1,22 +1,21 @@
-import {
-  NotificationProvider,
-  ThemeProvider,
-  UserProvider,
-} from "../context/provider";
-import { ItemsProvider } from "./ItemsProvider";
+import { ItemsProvider } from "./providers/items/ItemsProvider";
+import { NotificationProvider } from "./providers/notification/NotificationProvider";
+import { ThemeProvider } from "./providers/theme/ThemeProvider";
+import { UserProvider } from "./providers/user/UserProvider";
 
 type AppProviderProps = {
   children: React.ReactNode;
 };
 
+const providers = [
+  NotificationProvider,
+  ThemeProvider,
+  UserProvider,
+  ItemsProvider,
+];
+
 export function AppProvider({ children }: AppProviderProps) {
-  return (
-    <NotificationProvider>
-      <ThemeProvider>
-        <UserProvider>
-          <ItemsProvider>{children}</ItemsProvider>
-        </UserProvider>
-      </ThemeProvider>
-    </NotificationProvider>
-  );
+  return providers.reduceRight((acc, Provider) => {
+    return <Provider>{acc}</Provider>;
+  }, children);
 }
