--- packages/assignment/src/context/NotificationProvider.tsx
@@ -1,95 +1,0 @@
-import { createContext, useState } from "react";
-import { useCallback, useMemo } from "../@lib";
-import { renderLog } from "../utils";
-import { useNotificationContext } from "./hooks";
-
-type AlertType = "info" | "success" | "warning" | "error";
-
-interface Notification {
-  id: number;
-  message: string;
-  type: AlertType;
-}
-
-interface NotificationContextType {
-  notifications: Notification[];
-  addNotification: (message: string, type: Notification["type"]) => void;
-  removeNotification: (id: number) => void;
-}
-
-export const NotificationContext = createContext<
-  NotificationContextType | undefined
->(undefined);
-
-type ThemeProviderProps = {
-  children: React.ReactNode;
-};
-
-export const NotificationProvider = ({ children }: ThemeProviderProps) => {
-  const [notifications, setNotifications] = useState<
-    NotificationContextType["notifications"]
-  >([]);
-
-  const addNotification = useCallback(
-    (message: string, type: Notification["type"]) => {
-      const newNotification: Notification = {
-        id: Date.now(),
-        message,
-        type,
-      };
-      setNotifications((prev) => [...prev, newNotification]);
-    },
-    []
-  );
-
-  const removeNotification = useCallback((id: number) => {
-    setNotifications((prev) =>
-      prev.filter((notification) => notification.id !== id)
-    );
-  }, []);
-
-  const values = useMemo(
-    () => ({ notifications, addNotification, removeNotification }),
-    [notifications, addNotification, removeNotification]
-  );
-
-  return (
-    <NotificationContext.Provider value={values}>
-      {children}
-      <NotificationSystem />
-    </NotificationContext.Provider>
-  );
-};
-
-// NotificationSystem 컴포넌트
-const NotificationSystem: React.FC = () => {
-  renderLog("NotificationSystem rendered");
-  const { notifications, removeNotification } = useNotificationContext();
-
-  return (
-    <div className="fixed bottom-4 right-4 space-y-2">
-      {notifications.map((notification) => (
-        <div
-          key={notification.id}
-          className={`p-4 rounded shadow-lg ${
-            notification.type === "success"
-              ? "bg-green-500"
-              : notification.type === "error"
-              ? "bg-red-500"
-              : notification.type === "warning"
-              ? "bg-yellow-500"
-              : "bg-blue-500"
-          } text-white`}
-        >
-          {notification.message}
-          <button
-            onClick={() => removeNotification(notification.id)}
-            className="ml-4 text-white hover:text-gray-200"
-          >
-            닫기
-          </button>
-        </div>
-      ))}
-    </div>
-  );
-};
