--- packages/assignment/src/context/provider.ts
@@ -1,5 +1,0 @@
-import { ThemeProvider } from "./ThemeProvider";
-import { UserProvider } from "./UserProvider";
-import { NotificationProvider } from "./NotificationProvider";
-
-export { ThemeProvider, UserProvider, NotificationProvider };
