+++ packages/assignment/src/context/providers/theme/ThemeContext.ts
@@ -1,0 +1,12 @@
+import { createContext } from "react";
+
+type ThemeType = "light" | "dark";
+
+export interface ThemeContextType {
+  theme: ThemeType;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined
+);
