+++ packages/assignment/src/context/providers/user/UserContext.ts
@@ -1,0 +1,17 @@
+import { createContext } from "react";
+
+interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+export interface UserContextType {
+  user: User | null;
+  login: (email: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserContextType | undefined>(
+  undefined
+);
